refactor: Removed unused registry code
