+++ src/registries/ConfigRegistry.ts
