--- src/registries/NewRegistry/index.ts
@@ -1,3 +1,0 @@
-export * from './BlueBaseModuleRegistry';
-export * from './ComponentRegistry';
-export * from './Registry';
