+++ src/registries/PluginRegistry.ts
@@ -3,9 +3,9 @@ import {
 	BlueBaseModuleRegistryInputItem,
 	BlueBaseModuleRegistryItem,
 } from './BlueBaseModuleRegistry';
-import { MaybeThunk, getDefiniteBlueBaseModule, isBlueBaseModule } from '../../utils';
+import { MaybeThunk, getDefiniteBlueBaseModule, isBlueBaseModule } from '../utils';
 import { ComponentInputCollection } from './ComponentRegistry';
-import { DynamicIconProps } from '../../components';
+import { DynamicIconProps } from '../components';
 import { ItemCollection } from './Registry';
