+++ src/registries/__stories__/ComponentRegistry.stories.tsx
