+++ src/registries/__tests__/ConfigRegistry.test.ts
@@ -1,5 +1,5 @@
 // tslint:disable:object-literal-sort-keys
-import { BlueBase } from '../../../BlueBase';
+import { BlueBase } from '../../BlueBase';
 import { ConfigRegistry } from '../ConfigRegistry';
 
 const configs = {
