+++ src/registries/index.ts
@@ -1,6 +1,6 @@
-export * from './NewRegistry/ComponentRegistry';
-export * from './NewRegistry/ConfigRegistry';
-export * from './NewRegistry/HookRegistry';
-export * from './NewRegistry/PluginRegistry';
-export * from './NewRegistry/ThemeRegistry';
+export * from './ComponentRegistry';
+export * from './ConfigRegistry';
+export * from './HookRegistry';
+export * from './PluginRegistry';
+export * from './ThemeRegistry';
 export * from './Registry';
